8 refactoring on users client

+++ LetsDoIt.Moody/src/LetsDoIt.Moody.Persistence/Entities/CategoryDetail.cs
@@ -7,7 +7,7 @@ namespace LetsDoIt.Moody.Persistence.Entities
     {
         public int Id { get; set; }
         public int CategoryId { get; set; }
-        public int Order { get; set; }
+        public decimal Order { get; set; }
         public byte[] Image { get; set; }
         public DateTime CreatedDate { get; set; }
         public int CreatedBy { get; set; }
